backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic>=2.5.0
python-dotenv==1.0.0

# Database & Storage
sqlalchemy==2.0.23
chromadb==0.4.21
faiss-cpu==1.7.4

# LLM & Embeddings
langchain==0.1.7
ollama==0.1.0
openai>=1.6.1,<2
sentence-transformers>=2.2.2
huggingface-hub==0.19.4

# Text Processing
PyPDF2==3.0.1
python-docx>=0.8.11
markdown==3.5.1
unidecode==1.3.0

# Retrieval & Search
rank-bm25==0.2.2
numpy>=1.26,<2
scipy==1.11.4

# Evaluation
ragas==0.1.6
rouge-score==0.1.2
nltk==3.8.1

# Authentication & Security
pyjwt>=2.8,<3
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utilities
pandas==2.1.3
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
